created seeds for questions & answers tables
