+++ src/Routes/Users/AccountRouter.js
@@ -54,62 +54,4 @@ accountRouter.post('/', jsonBodyParser, async (req, res, next) => {
   }
 });
 
-accountRouter.post('/initial', jsonBodyParser, (req, res, next) => {
-  const {
-    user_id,
-    birthdate,
-    location,
-    nationality,
-    gender,
-    college_graduate
-  } = req.body;
-
-  if (!user_id) {
-    logger.error('User ID is required');
-    return res.status(400).send('User ID required');
-  }
-  if (!birthdate) {
-    return res.status(400).send('Birthdate required');
-  }
-  if (!nationality) {
-    return res.status(400).send('Nationality required');
-  }
-  if (!gender) {
-    return res.status(400).send('Gender required');
-  }
-  if (!college_graduate) {
-    return res.status(400).send('CollegeGraduate required');
-  }
-  if (!location) {
-    return res.status(400).send('Location required');
-  }
-
-  const initialAnswers = {
-    user_id,
-    birthdate,
-    location,
-    nationality,
-    gender,
-    college_graduate
-  };
-
-  AccountService.insertAnswers(req.app.get('db'), initialAnswers)
-    .then(result => res.status(201).json(result))
-    .catch(next);
-});
-
-accountRouter.get('/initial/:user_id', (req, res) => {
-  const { user_id } = req.params;
-
-  // grabs the user_id from the params then searchs the users_info table to see if theyve completed the initial quiz
-  // if they have the api sends 'true' back to the user and 'false' if not
-  AccountService.initialStatus(req.app.get('db'), user_id).then(result => {
-    if (result !== undefined) {
-      return res.status(200).json(true);
-    } else {
-      return res.status(200).json(false);
-    }
-  });
-});
-
 module.exports = accountRouter;
