+++ src/Routes/Users/AccountService.js
@@ -43,22 +43,6 @@ const AccountService = {
       last_name: user.last_name,
       username: user.username
     };
-  },
-  insertAnswers(db, answers) {
-    return db
-      .into('users_info')
-      .insert(answers)
-      .returning('*')
-      .then(rows => {
-        return rows[0];
-      });
-  },
-  initialStatus(db, user_id) {
-    return db
-      .from('users_info')
-      .where({ user_id })
-      .first()
-      .returning('*');
   }
 };
